+++ minecraft/src/net/minecraft/client/gui/screens/inventory/HangingSignEditScreen.java
@@ -2,7 +2,6 @@ package net.minecraft.client.gui.screens.inventory;
 
 import com.mojang.blaze3d.systems.RenderSystem;
 import com.mojang.blaze3d.vertex.PoseStack;
-import com.mojang.math.Vector3f;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.minecraft.client.renderer.MultiBufferSource;
@@ -10,6 +9,7 @@ import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.level.block.entity.SignBlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
+import org.joml.Vector3f;
 
 @Environment(EnvType.CLIENT)
 public class HangingSignEditScreen extends AbstractSignEditScreen {
@@ -25,12 +25,12 @@ public class HangingSignEditScreen extends AbstractSignEditScreen {
 
 	@Override
 	protected void offsetSign(PoseStack poseStack, BlockState blockState) {
-		poseStack.translate((double)((float)this.width / 2.0F), 125.0, 50.0);
+		poseStack.translate((float)this.width / 2.0F, 125.0F, 50.0F);
 	}
 
 	@Override
 	protected void renderSignBackground(PoseStack poseStack, MultiBufferSource.BufferSource bufferSource, BlockState blockState) {
-		poseStack.translate(0.0, -13.0, 0.0);
+		poseStack.translate(0.0F, -13.0F, 0.0F);
 		RenderSystem.setShaderTexture(0, this.texture);
 		RenderSystem.setShaderColor(1.0F, 1.0F, 1.0F, 1.0F);
 		poseStack.scale(4.0F, 4.0F, 1.0F);
