+++ minecraft/src/net/minecraft/client/renderer/blockentity/SkullBlockRenderer.java
@@ -85,10 +85,10 @@ public class SkullBlockRenderer implements BlockEntityRenderer<SkullBlockEntity>
 	) {
 		poseStack.pushPose();
 		if (direction == null) {
-			poseStack.translate(0.5, 0.0, 0.5);
+			poseStack.translate(0.5F, 0.0F, 0.5F);
 		} else {
 			float h = 0.25F;
-			poseStack.translate((double)(0.5F - (float)direction.getStepX() * 0.25F), 0.25, (double)(0.5F - (float)direction.getStepZ() * 0.25F));
+			poseStack.translate(0.5F - (float)direction.getStepX() * 0.25F, 0.25F, 0.5F - (float)direction.getStepZ() * 0.25F);
 		}
 
 		poseStack.scale(-1.0F, -1.0F, 1.0F);
